requirements.txt
pandas>=2.0
requests>=2.31
pyarrow>=12.0
PyYAML>=6.0
